requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
chromadb
openai
pypdf
tiktoken
google-generativeai
google-genai
python-multipart
pymongo
sentence-transformers
